Rust: Updates for SDK changes Oct 30 2023 (#5588)
* Updates for Rust changes Oct 30 2023
* Use rt-tokio in aws-smithy-types for aws_sdk_s3::primitives::ByteStream::from_path_body_0_4
* Bump READMEs

+++ rust_dev_preview/run_all.sh
@@ -1,4 +1,4 @@
-#!/usr/bin/env bash
+#!/usr/bin/env bash -x
 
 cd "$(dirname "$0")"
 
@@ -10,13 +10,15 @@ DIRS=(
 )
 
 ACTIONS=(
-  "fmt --check"
-  clippy
-  test
+  "fmt --check --all"
+  "clippy --all"
+  "test --all"
 )
 
-# Clean Cargo.log
-rm **/Cargo.lock
+if [ "$1" == "--clean" ] ; then 
+  rm **/Cargo.lock
+  ACTIONS=("clean" "${ACTIONS[@]}")
+fi
 
 export RUSTFLAGS="-D warnings" ;
 export APP_ENVIRONMENT="test"
@@ -26,10 +28,10 @@ CARGO="$HOME/.cargo/bin/cargo"
 FAIL=()
 for f in "${DIRS[@]}" ; do
   for a in "${ACTIONS[@]}" ; do
-    "$CARGO" $a --manifest-path $f/Cargo.toml --all || FAIL+=("${a}:$f")
+    "$CARGO" $a --manifest-path $f/Cargo.toml || FAIL+=("${a}:$f")
   done
 done
 
-echo ${FAIL[@]}
+echo "${FAIL[@]}"
 exit ${#FAIL[@]}
